Various minor bugfixes.
(cherry picked from commit 8d3e28aaf0727647d75f39473ce609ea319cb9a7)

+++ src/core/input/winInput.cpp
@@ -371,7 +371,7 @@ BOOL CALLBACK initJoystick(const DIDEVICEINSTANCE* instance, VOID* _context)
     //}
 
     // Register Joystick
-    input->joysticks.push_back(std::make_shared<WinJoystick>(input->joysticks.size(), handle));
+    input->joysticks.push_back(std::make_shared<WinJoystick>(static_cast<uint32>(input->joysticks.size()), handle));
     return DIENUM_CONTINUE;
 }
