+++ src/core/rendering/vulkan/vkDevice.h
@@ -173,7 +173,7 @@ class VulkanDevice : public CommonDevice
 
     virtual std::shared_ptr<Shader>  createShader(const ShaderStage stage,
                                              const uint8* data,
-                                             const uint64 size);
+                                             const uint32 size);
 
     virtual Pipeline* createPipeline(
         const PipelineState& pipelineState);
